requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-arango>=7.8.0
pandas>=2.2.2
numpy>=2.1.0
requests>=2.32.3
beautifulsoup4==4.12.0
pdfplumber==0.11.0
python-dotenv==1.0.0
